+++ Module.php
@@ -2,8 +2,9 @@
 
 namespace humhub\modules\scrollup;
 
+use humhub\components\Module as BaseModule;
 
-class Module extends \humhub\components\Module
+class Module extends BaseModule
 {
 
     public function getPosition()
@@ -13,5 +14,14 @@ class Module extends \humhub\components\Module
             return '';
         }
         return $position;
+    }
+
+    public function getColor()
+    {
+        $color = $this->settings->get('color');
+        if (empty($color)) {
+            return '';
+        }
+        return $color;
     }
 }